patch to compile on centos6, TODO align data types to stdint

+++ chaos/common/data/CDataWrapper.cpp
@@ -107,7 +107,8 @@ void CDataWrapper::appendInt64ToArray(int64_t int64ArrayElement) {
 
 //append a strin gto an open array
 void CDataWrapper::appendDoubleToArray(double_t doubleArrayElement) {
-    bsonArrayBuilder->append(doubleArrayElement);
+  // TODO: all std int or NOT
+  bsonArrayBuilder->append((double)doubleArrayElement);
 }
 
 //appen a CDataWrapper to an open array
